requirements.txt
python-dotenv==1.0.0
pyTelegramBotAPI==4.11.1
gspread==5.7.0
oauth2client==4.1.3